+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntityModelTest.java
@@ -67,7 +67,21 @@ public class DefaultEntityModelTest extends AbstractEntityModelTest<DefaultEntit
 	public static final class TestEntityEditor extends AbstractEntityEditor<TestEntityEditor> {
 
 		public TestEntityEditor(EntityType entityType, EntityConnectionProvider connectionProvider) {
-			super(entityType, connectionProvider, new TestComponentModels() {});
+			this(entityType, connectionProvider, new TestComponentModels() {});
+		}
+
+		public TestEntityEditor(EntityType entityType, EntityConnectionProvider connectionProvider, TestComponentModels componentModels) {
+			super(entityType, connectionProvider, componentModels);
+		}
+
+		@Override
+		public TestEntityEditor create(EntityType entityType) {
+			return new TestEntityEditor(entityType, connectionProvider());
+		}
+
+		@Override
+		public TestEntityEditor create(EntityType entityType, ComponentModels componentModels) {
+			return new TestEntityEditor(entityType, connectionProvider(), (TestComponentModels) componentModels);
 		}
 	}